requirements.txt
flask
pyrebase4
Flask-Session
python-dotenv
flask-sqlalchemy
psycopg2-binary
gunicorn